ViewItemFX: Progress bar fix
The progress bar 'setProgress' method expects a double value  (0.0 -
1.0) but currently had an int so no progress was showing. The value has
been cast to double to fix this.

+++ src/ViewItemFX.java
@@ -44,6 +44,6 @@ public class ViewItemFX extends StackPane
 			return;
 		}
 		
-		progress.setProgress(value/total);
+		progress.setProgress((double)value/total);
 	}
 }
